requirements.txt
streamlit
requests
websocket-client
streamlit-drawable-canvas
Pillow